requirements.txt
torchvision
numpy
pillow
scikit-learn
matplotlib
tqdm
ipdb
pytube
moviepy